get median of 3 working with param
