+++ src/algorithms/controllers/quickSortM3.js
@@ -1,25 +1,9 @@
-// import 1D tracer to generate array in a separate component of the middle panel
-import ArrayTracer from '../../components/DataStructures/Array/Array1DTracer';
+import { initVisualisers , run_QS, return_qs } from './quickSort_shared'
+
+const is_qs_median_of_3 = true
+const run = run_QS(is_qs_median_of_3)
 
 export default {
-
-  initVisualisers() {
-    return {
-      array: {
-        instance: new ArrayTracer('array', null, 'Array view', {
-          arrayItemMagnitudes: true,
-        }), // Label the input array as array view
-        order: 0,
-      },
-    };
-  },
-
-  /**
-   *
-   * @param {object} chunker
-   * @param {array} nodes array of numbers needs to be sorted
-   */
-  run(chunker, { nodes }) {
-    chunker.add(1)
-  },
+    initVisualisers, 
+    run
 };
